Rename and move test
